Disable tables when converting markdown to draft
DraftJS doesn’t support tables (at least not out of the box) and if we
have them turned on in markdown it creates missing blocks of text where
the table content is.

Since someone could theoretically write a DraftJS plugin to support
tables (in fact I think there is one already!) We may want to make this
an option to turn on at some point so people can write their own
entityhandler for tables in markdownToDraft/draftToMarkdown, but I
wasn’t sure what the best API for that would be so it’s just a hardcoded
value for now.